+++ src/SFML/Network/Win32/SocketImpl.cpp
@@ -34,7 +34,7 @@ namespace sf
 namespace priv
 {
 ////////////////////////////////////////////////////////////
-sockaddr_in SocketImpl::CreateAddress(unsigned long address, unsigned short port)
+sockaddr_in SocketImpl::CreateAddress(Uint32 address, unsigned short port)
 {
     sockaddr_in addr;
     std::memset(addr.sin_zero, 0, sizeof(addr.sin_zero));
@@ -63,7 +63,7 @@ void SocketImpl::Close(SocketHandle sock)
 ////////////////////////////////////////////////////////////
 void SocketImpl::SetBlocking(SocketHandle sock, bool block)
 {
-    unsigned long blocking = block ? 0 : 1;
+    u_long blocking = block ? 0 : 1;
     ioctlsocket(sock, FIONBIO, &blocking);
 }
